+++ src/main/java/com/github/jaksonlin/testcraft/util/ClassFileInfo.java
@@ -1,14 +1,17 @@
 package com.github.jaksonlin.testcraft.util;
 
+import java.util.List;
 public class ClassFileInfo {
     private final String fullyQualifiedName;
     private final String className;
     private final String packageName;
+    private final List<String> methods;
 
-    public ClassFileInfo(String fullyQualifiedName, String className, String packageName) {
+    public ClassFileInfo(String fullyQualifiedName, String className, String packageName, List<String> methods) {
         this.fullyQualifiedName = fullyQualifiedName;
         this.className = className;
         this.packageName = packageName;
+        this.methods = methods;
     }
 
     public String getFullyQualifiedName() {
@@ -21,5 +24,9 @@ public class ClassFileInfo {
 
     public String getPackageName() {
         return packageName;
+    }
+
+    public List<String> getMethods() {
+        return methods;
     }
 }